refactor(config): reduce public facade
